SwaMe.Test folder name changes
Also the tsv file that is no longer used was deleted.
