--- SwaMeTest/SwaMeTest.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29306.81
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SwaMeTest", "SwaMeTest\SwaMeTest.csproj", "{4893652C-E803-4748-939B-F29AE9D08ADC}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{4893652C-E803-4748-939B-F29AE9D08ADC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4893652C-E803-4748-939B-F29AE9D08ADC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4893652C-E803-4748-939B-F29AE9D08ADC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4893652C-E803-4748-939B-F29AE9D08ADC}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {7B73C04D-4C13-4C99-8CBE-B774DCBF1C32}
-	EndGlobalSection
-EndGlobal
